refactor(viewer): modify calculateLengthBetweenPoints location
