+++ libs/insight-viewer/src/Annotation/utils/calculateLengthBetweenPoints.spec.ts
