+++ libs/insight-viewer/src/Annotation/utils/calculateLengthBetweenPoints.ts
