+++ libs/insight-viewer/src/Annotation/utils/getLineLength.ts
@@ -1,7 +1,7 @@
 import { Point } from '../../types'
 import { Image } from '../../Viewer/types'
-import { calculateLengthBetweenPoints } from '../../utils/common/calculateLengthBetweenPoints'
 import { IMAGER_PIXEL_SPACING } from '../../utils/common/const'
+import { calculateLengthBetweenPoints } from './calculateLengthBetweenPoints'
 
 export function getLineLength(
   startPoint: Point,
